Make changes from Cataloging_notes_v.6
_From Cataloging\_notes\_v.6_ -- some things have been addressed:

> Title group -- move to below Comments

Change in place, and will be on the production server later tonight.

> Sex - add "unknown" to controlled static list

I can't get this configuration to take effect; will need to ask the Talk
list.

> Production Culture -- still not linking to the authority list (no gray
> arrow in field)

I can't get this configuration to take effect; will need to ask the Talk
list.

> Associated Cultural Affinity - still not linking to the authority list
> (no gray arrow in field)

Change in place, and will be on the production server later tonight.

> Label Credits - title still shows as "Credits" and is not yet
> populating with PP data.
>

Label change in place, and will be on the production server later
tonight.